renamed keyboard_input.js and render_tasks.js for consistency
